Moved to library file.
